Add an envvar to permit valgrinding of a particular nspluginviewer because: 1) It's needed sometimes 2) Enabling is globally can result in unpleasant experiences of having your machine nearly DoS'd by a bunch of instances popping up due to website's gratious flash use, while you're trying to look something up quickly while on the phone.
CCBUG: 241284

svn path=/trunk/KDE/kdebase/apps/; revision=1136791

+++ nsplugins/nspluginloader.cpp
@@ -342,6 +342,16 @@ bool NSPluginLoader::loadViewer()
       return false;
    }
 
+   bool runInValgrind = !qgetenv("VALGRIND_NSPLUGINVIEWER").isEmpty();
+
+   if (runInValgrind) {
+      _process << "konsole";
+      _process << "--noclose";      
+      _process << "-e";
+      _process << "valgrind";
+      _process << "--num-callers=50";
+   }
+
    _process << viewer;
 
    _process << "-dbusservice";
@@ -358,7 +368,7 @@ bool NSPluginLoader::loadViewer()
       //kapp->processEvents(); // would lead to recursive calls in khtml
       usleep( 50*1000 );
       cnt++;
-      if (cnt >= 100)
+      if ( cnt >= 100 && !runInValgrind )
       {
          kDebug() << "timeout";
          _process.kill();
